% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cna.R
\docType{data}
\name{cna}
\alias{cna}
\alias{example_cna}
\title{Copy number alterations for sample data}
\format{
## `cna`
a dataframe with 19243 rows and 4 columns
\describe{
 \item{Hugo_Symbol}{Hugo Symbol Gene ID}
 \item{Sample1}{Copy number threshold calls for example sample 1}
 \item{Sample2}{Copy number threshold calls for example sample 2}
 \item{Sample3}{Copy number threshold calls for example sample 3}
}
}
\usage{
cna
}
\value{
A dataframe containing copy
number alterations represented with GISTIC gene-level copy number
threshold calls for three samples
}
\description{
A dataframe containing copy
number alterations represented with GISTIC gene-level copy number
threshold calls for three samples
}
\details{
cna
}
\keyword{datasets}
