% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_models.R
\name{download_models}
\alias{download_models}
\title{Download pre-trained models from OSF}
\usage{
download_models(pathways = NULL, keyword = NULL, keep_files = FALSE)
}
\arguments{
\item{pathways}{a list of pathways}

\item{keyword}{a character key word for searching pathways}

\item{keep_files}{logical value to specify removal of downloaded files}
}
\value{
a list of pretrained models
}
\description{
This function downloads the models pre-trained on TCGA pan-cancer datasets
for calculating scores of 47 different pathways using genomic data.
}
\details{
RDS files are downloaded to current working directory. Downloaded files
may optionally be removed from directory after loaded into R environment.
}
\examples{
my_pathways <- c("HALLMARK_DNA_REPAIR", "HALLMARK_G2M_CHECKPOINT")
my_models <- download_models(pathways = my_pathways)

my_models <- download_models(keyword = "MYC")

}
