% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Curate datasets to use with PhenoMap models}
\usage{
format_data(snv, cna)
}
\arguments{
\item{snv}{a dataframe including single nucleotide variants in vcf format
containing the following columns:
Hugo_Symbol
Tumor_Sample_Barcode}

\item{cna}{a dataframe including copy number alterations represented
with GISTIC gene-level copy number threshold calls}
}
\value{
a dataframe containing merged genomic data
}
\description{
This function take single nucleotide variants in vcf format with copy
number alterations represented with GISTIC gene-level copy number
threshold calls.
}
\details{
The two datasets are merged and formatted for use with
PhenoMap models. Sample names between two dataframes should match. If a
sample name appears in the cna file, but not the snv file, it will be
assumed this sample had no vaiant detected.
}
\examples{
data("example_cna")
data("example_snv")
data <- format_data(snv = snv, cna = cna)

}
