% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_df.R
\docType{data}
\name{gene_df}
\alias{gene_df}
\alias{genes_for_models}
\title{Required gene inputs for model}
\format{
## `gene_df`
a dataframe with 6114 rows and 1 column
\describe{
 \item{Gene}{List of gene names used in models}
}
}
\usage{
gene_df
}
\value{
A dataframe containing list of genes
}
\description{
A dataframe containing list of genes required for model input used
for data formatting.
}
\details{
gene_df
}
\keyword{datasets}
