% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_descriptions.R
\docType{data}
\name{pathway_descriptions}
\alias{pathway_descriptions}
\title{Descriptions of available models for download from OSF}
\format{
## `pathway_descriptions`
a dataframe with 2 columns and 1545 rows
\describe{
 \item{Pathway}{Name of pathway}
 \item{MSigDB.Collection}{MSigDB Collection}
}
}
\usage{
pathway_descriptions
}
\value{
A dataframe containing list of model names and MSigDB collections
}
\description{
A dataframe containing list of model names and MSigDB collections
}
\details{
pathway_descriptions
}
\keyword{datasets}
