% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_files.R
\docType{data}
\name{pathway_files}
\alias{pathway_files}
\title{Locations of available models for download from OSF}
\format{
## `pathway_files`
a dataframe with 3 columns and 1545 rows
\describe{
 \item{name}{Name of file}
 \item{id}{File ID for OSF}
 \item{meta}{Metadata for locating file on OSF repository}
}
}
\usage{
pathway_files
}
\value{
A dataframe containing list of model names and locations to
download from OSF
}
\description{
A dataframe containing list of model names and locations to
download from OSF
}
\details{
pathway_files
}
\keyword{datasets}
