% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoMapper.R
\name{PhenoMapper}
\alias{PhenoMapper}
\title{PhenoMapper: single sample pathway scores for genomic data}
\format{
a software package with four functions and 5 data files
}
\usage{
PhenoMapper
}
\value{
a software packages for downloading pretrained models to calculate
single sample pathway scores from genomic datasets
}
\description{
A software package for downloading pretrained models to calculate
single sample pathway scores from genomic datasets
}
\details{
PhenoMapper
}
\keyword{packages}
