% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_scores.R
\name{predict_scores}
\alias{predict_scores}
\title{Calculate GSVA-based pathway scores with genomic data}
\usage{
predict_scores(data, models)
}
\arguments{
\item{data}{a dataframe of formatted snv and cna datasets}

\item{models}{a list of models downloaded from OSF}
}
\value{
a dataframe containing single sample scores for each pathway
}
\description{
This function uses our models pretrained on TCGA pan-cancer datasets
to calculate single sample pathway-level scores based on single nucleotide
variant and copy number alteration data.
}
\examples{
data("example_cna")
data("example_snv")

data <- format_data(snv = snv, cna = cna)

my_pathways <- c("HALLMARK_DNA_REPAIR", "HALLMARK_G2M_CHECKPOINT")
my_models <- download_models(pathways = my_pathways, keep_files = FALSE)

my_scores <- predict_scores(data = data, models = my_models)

}
