% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv.R
\docType{data}
\name{snv}
\alias{snv}
\alias{example_snv}
\title{Single nucleotide variants for sample data}
\format{
## `snv`
a dataframe with 36 columns and 1570 rows
\describe{
 \item{Hugo_Symbol}{Hugo Symbol Gene ID}
 \item{Tumor_Sample_Barcode}{Sample ID}
}
}
\usage{
snv
}
\value{
A dataframe containing vcf file of three samples
}
\description{
dataframe containing the vcf format file of single nucleotide
variants from three samples
}
\details{
snv
}
\keyword{datasets}
