% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_models.R
\name{view_models}
\alias{view_models}
\title{View pre-trained models available on OSF}
\usage{
view_models(keyword = "ALL")
}
\arguments{
\item{keyword}{a character key word for searching pathways}
}
\value{
a list of model names
}
\description{
This function allows users to browse available models for download
}
\examples{
all_pathways <- view_models()
kegg_pathways <- view_models(keyword = "KEGG")

}
